Reader for LeetCodeRatio.csv
Reads in the information from LeetCodeRatio.csv and writes to popup.html. HTML has a table with the question names, difficulty, likes, dislikes, and like-dislike ratio. Question names are clickable and link to the question on LeetCode. Difficulty is color coded to match LeetCode's format. Like-dislike ratio is just a percentage right now.